Make typing-extensions a runtime dependency
The type declarations here need to be imported at runtime. There doesn't
seem to be an easy way to tell Python to ignore any errors if the type
declarations can't be parsed properly. There is deferred evaluation of
annotation which, when very carefully used may allow this, but that is
only available in 3.7 and higher, while we support 3.6 and higher.